refactor: Extract e2e utils to the proper file
